move utils from scripts to moodle_export
