+++ moodle_export/grades_parser.py
@@ -168,8 +168,8 @@ class Main:
                 # write data to yandex disk
                 if cls.args.yandex_token and cls.args.yandex_path:
                     # TODO: refactor нadisk
-                    from ..scripts import utils
-                    utils.write_sheet_to_file(cls.args.yandex_path, csv_path, sheet_name="Онлайн-курс")
+                    from utils import write_sheet_to_file
+                    write_sheet_to_file(cls.args.yandex_path, csv_path, sheet_name="Онлайн-курс")
                     
                     yandex_path = cls.args.yandex_path
                     print(f'Course {cls.args.course_id} uploaded to table on Disk! Path to the table is: {yandex_path}')
